[mod] display template tile
